random bug fixed

+++ src/main/java/hu/bme/mit/spaceship/TorpedoStore.java
@@ -37,7 +37,7 @@ public class TorpedoStore {
     boolean success = false;
     // simulate random overheating of the launcher bay which prevents firing
     Random generator = SecureRandom.getInstanceStrong();
-    double r = generator.nextDouble();
+    double r = generator.nextInt();
 
     if (r >= FAILURE_RATE) {
       // successful firing
